# chore(print): improve performance for data loading