--- a/frontend/src/components/print/config/ActivityConfig.vue
+++ b/frontend/src/components/print/config/ActivityConfig.vue
@@ -8,7 +8,7 @@
 
 <script>
 export default {
-  name: 'Activity',
+  name: 'ActivityConfig',
   props: {
     value: { type: Object, required: true },
     camp: { type: Object, required: true }