--- a/frontend/src/components/print/config/CoverConfig.vue
+++ b/frontend/src/components/print/config/CoverConfig.vue
@@ -4,7 +4,7 @@
 
 <script>
 export default {
-  name: 'CoverConfig',
+  name: 'Cover',
   props: {
     value: { type: Object, required: true }
   },