--- a/frontend/src/components/print/config/PicassoConfig.vue
+++ b/frontend/src/components/print/config/PicassoConfig.vue
@@ -12,7 +12,7 @@
 
 <script>
 export default {
-  name: 'PicassoConfig',
+  name: 'Picasso',
   props: {
     value: { type: Object, required: true },
     camp: { type: Object, required: true }