--- a/frontend/src/components/print/config/PicassoConfig.vue
+++ b/frontend/src/components/print/config/PicassoConfig.vue
@@ -12,7 +12,7 @@
 
 <script>
 export default {
-  name: 'Picasso',
+  name: 'PicassoConfig',
   props: {
     value: { type: Object, required: true },
     camp: { type: Object, required: true }