--- a/print/assets/tailwind.css
+++ b/print/assets/tailwind.css
@@ -1,0 +1,1 @@
+/* global css styling */