… api/tests/Api/ContentNodes/ColumnLayout/ListColumnLayoutTest.php
… api/tests/Api/ContentNodes/ColumnLayout/ReadColumnLayoutTest.php
… api/tests/Api/ContentNodes/ContentNode/ListContentNodesTest.php
… api/tests/Api/ContentNodes/ContentNode/ReadContentNodeTest.php
… api/tests/Api/ContentNodes/CreateContentNodeTestCase.php
… api/tests/Api/ContentNodes/MaterialNode/ListMaterialNodeTest.php
… api/tests/Api/ContentNodes/MultiSelect/ListMultiSelectTest.php
… api/tests/Api/ContentNodes/SingleText/ListSingleTextTest.php
… api/tests/Api/ContentNodes/Storyboard/ListStoryboardTest.php
… api/tests/Doctrine/Filter/ContentNodePeriodFilterTest.php
… common/locales/de.json
… common/locales/en.json
… frontend/src/components/print/print-react/PrintPreviewReact.vue
… frontend/src/components/print/print-react/components/RichText.jsx
… frontend/src/components/print/print-react/documents/pdfDocument/Component.jsx
… frontend/src/components/print/print-react/documents/pdfDocument/prepareInMainThread.js
… frontend/src/components/print/print-react/generatePdf.js
… frontend/src/components/story/StoryDay.vue
… frontend/src/locales/de-CH-scout.json
… frontend/src/locales/de.json
… frontend/src/locales/en.json
… frontend/src/locales/fr.json
… frontend/src/locales/it.json
… frontend/src/router.js
… frontend/src/views/activity/Activity.vue
… frontend/src/views/camp/CampProgram.vue
… frontend/src/views/camp/Print.vue
… frontend/src/views/camp/PrintActivity.vue
… frontend/src/views/camp/PrintCamp.vue
… frontend/src/views/camp/PrintPeriod.vue